Send target direction to clients

+++ server/src/main/java/server/EventSocket.java
@@ -23,7 +23,7 @@ public class EventSocket {
 
     @OnMessage
     public void onWebSocketMessage(Session session, ByteBuffer buffer) {
-        double alpha = buffer.getDouble(0);
+        float alpha = (float) buffer.getDouble(0); // TODO
         boolean fast = buffer.get(8) != 0;
         assert(buffer.get(9) == 42);
         SnakeServer.updateUserInput(session, alpha, fast);
